+++ test/synapse_build/test_sample.py
@@ -1,8 +1,0 @@
-
-import pytest
-
-from src.synapse_build.main import sample
-
-def test_sample():
-  res = sample()
-  assert res == 'Hello World'