+++ docs/CHANGELOG.md
@@ -1,6 +1,21 @@
 # Release Notes 
 ## Pull Requests
-**1** Feature/repo_setup
+**3** Added core models and build client for synapse build
 ### Commits
-  - **dc31e8831b6a6a344d0af5cbafa226957dc61c94** configured template repo
-  - **8cd780e9e7cbffde3e75286b19ed554660d53806** delete template docs
+  - **a0b70e67dd35cf00030e61e81d33e75c80b7a7f1** Merge branch &#x27;main&#x27; of https://github.com/BMeyn/synapse_build
+  - **466db14257ba4e7b6647988b3142af76ef4c4819** added tmp folder to gitignore
+  - **e7d2de70e1c1a75cc6770db19afdb097da83c85b** removed usused dependencies
+  - **ceab55df75bc956e9762f60a32b047e8ae96a33d** added vscdoe extensions to container config
+  - **ee553fe147b82382930ce5539fa16532b0287389** added launch.json config for vscode debugging
+  - **a57b54e6ed05eb4d7b6ff507e2d84af3cdef2802** added models for synapse build
+  - **dd331590563355e80b344d185996dd75e9c5ad49** added build client for synapse build
+  - **812c1af766116010a8fa7052c158c1374653ad2a** implemented tests for synapse build models and build
+  - **ac939ea21110c62b7b6c155367fb713363fc502f** added pytest cases dependencies
+  - **3b82a1c855712c14360bfa0cd75ff0ab8f4f4b44** added jsonschema to dependencies
+  - **bb35a21471dff1a9efb4d6180e64dbd9efad1b0d** debug parameterConfig tests
+  - **e5501ae61a6539fa1deeb427f9666dc323e63ce4** removed old files
+  - **9442973bea5f2e8f0e0f82c292be4980ce590abb** disable flake8 tests
+  - **b5a5d2382b3d81f4d1560a2d40384f5419c73b83** addded verbose to pytest output
+  - **d8c81eb0683aca6522bf42c5da0b4b3711e9f193** added utf8 encoding when data is save to file
+  - **3f1461ed58344f80f2660134222c427bda0ffb94** added sort_keys to json dumps
+  - **bde355f4b3d010cc1d44864ee204a8e6deeb9894** added step in test_buildClient to sort results by dict value